feat: add identity detail entry for identity card
